feat: added player stats at the end of the game, removed openjk & windows support, added zydis as static dependency, restored the original SDK, added a way to hook on jampgame too, changed lib name

+++ src/jampgame_proxy/Wrappers/Proxy_OriginalAPI_Wrappers.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "sdk/qcommon/q_platform.hpp"
+#include <sdk/qcommon/qcommon.h>
 
 // VM_DllSyscall can handle up to 1 (command) + 16 args
-intptr_t QDECL Proxy_OriginalAPI_VM_DllSyscall(intptr_t command, ...);
+int QDECL Proxy_OriginalAPI_VM_DllSyscall(int command, ...);
 // VM_Call can handle up to 1 (command) + 11 args
-intptr_t QDECL Proxy_OriginalAPI_VM_Call(intptr_t command, ...);
+int QDECL Proxy_OriginalAPI_VM_Call(int command, ...);
